coin toss experiment
This notebook illustrates the coin toss experiment where a certain number of coins(n) are tossed simultaneously and this experiment keeps on repeating again and again for N times. 
Here random variable, X=number of heads in 1 experiment of tossing n coins.
This is a case of Binomial Probability Distribution. 
Here we'll visualise how Probability Distribution is just and extended version of Relative Frequency Distribution as sample size limits to infinity.